+++ src/Plugin/Wechat/StartPlugin.php
@@ -10,7 +10,7 @@ use Yansongda\Artful\Contract\PluginInterface;
 use Yansongda\Artful\Logger;
 use Yansongda\Artful\Rocket;
 
-#[Deprecated(reason: '自 v3.7.5 版本已废弃，请直接使用 `yansongda/artful` 包中的 `Yansongda\Artful\Plugin\StartPlugin`')]
+#[Deprecated(reason: '自 v3.7.5 版本已废弃', replacement: '`yansongda/artful` 包中的 `Yansongda\Artful\Plugin\StartPlugin`')]
 class StartPlugin implements PluginInterface
 {
     public function assembly(Rocket $rocket, Closure $next): Rocket